fixing build

+++ src/cost_model_calibration_lib/feature/aggregate_features.cpp
@@ -5,8 +5,8 @@ namespace cost_model {
 
 const std::map<std::string, AllTypeVariant> AggregateFeatures::serialize() const {
   return {
-      {"aggregate_column_count", aggregate_column_count},
-      {"group_by_column_count", group_by_column_count},
+      {"aggregate_column_count", static_cast<int64_t>(aggregate_column_count)},
+      {"group_by_column_count", static_cast<int64_t>(group_by_column_count)},
   };
 }
